requirements.txt
asgiref==3.11.1
cloudinary==1.44.1
dj-database-url==3.1.0
Django==6.0.2
django-allauth==65.14.1
django-ckeditor-5==0.2.19
django-cloudinary-storage==0.3.0
django-jazzmin==3.0.2
django-js-asset==3.1.2
gunicorn==25.0.3
pillow==12.1.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
stripe==14.3.0
tzdata==2025.3
whitenoise==6.11.0
